blur-processor/requirements.txt
Flask==2.3.3

Pillow
redis